Update cars project structure for adding logistic regression
