+++ cars/src/logistic-regression/index.ts
@@ -1,0 +1,2 @@
+console.log('\nExecuting Logistic regression analysis 📈\n');
+console.log('\nFinished Logistic regression analysis 📉');
